% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grade_homework2}
\alias{grade_homework2}
\title{Function to grade 560R:Intro to R for Epi Homework 2}
\usage{
grade_homework2(file)
}
\arguments{
\item{file}{Name of file with extension (e.g. "Krall_560R_hw2.R")}
}
\description{
This function will provide general feedback for homework 2
}

