% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hw2_1_sleep}
\alias{hw2_1_sleep}
\title{Function to Part 1 in Homework 2}
\usage{
hw2_1_sleep(file)
}
\arguments{
\item{file}{Name of file with extension (e.g. "Krall_560R_hw2.R")}
}
\description{
This function will grade Homework 2 part 1
}

